refactor: separate types and interfaces to their own folders
