+++ .eslintrc.json
@@ -19,6 +19,7 @@
     "rules": {
         "no-debugger": "off",
         "no-console": 0,
-        "class-methods-use-this": "off"
+        "class-methods-use-this": "off",
+        "prettier/prettier": ["error", { "endOfLine": "auto" }]
     }
 }
