+++ src/components/interfaces/interfaces.ts
@@ -1,0 +1,9 @@
+import { NewsData, NewsSources } from '../types/types';
+
+export interface drawNewsFunction {
+    (data?: NewsData): void;
+}
+
+export interface drawSourcesFunction {
+    (data?: NewsSources): void;
+}
